pyenv/src/requirement.txt
mysql-connector-python
scrapy